requirements.txt
Flask >= 3.0, <= 3.1
brother-ql-inventree >= 1.3
Pillow == 10.*
treepoem
python-dotenv == 1.*
gunicorn
qrcode[pil] == 7.4.*